refactor(421): extract regex + html host families to vera/runtime/
Step 6 of the codegen/api.py decomposition (#421).  The Regex (§9.6.15) and HTML (§9.7.4) families move out of execute() into vera/runtime/regex.py and vera/runtime/html.py via register_regex/register_html.  Both are stateless (the apparent 'store' uses flagged earlier were all in comments) -- the host callbacks rely only on the now-complete vera.runtime.heap helpers plus their own lazy imports (re, html.parser, vera.wasm.html_serde).  html.py needed an explicit 'from typing import Any' (the annotation lived on api.py's module imports before).  api.py 2,592->2,171 lines.  mypy + ruff clean; 40 regex/html codegen tests pass and ch09_regex/ch09_html run end-to-end.
